+++ platform-api/src/main/java/com/flow/platform/api/dao/ArtifactDaoImpl.java
@@ -16,6 +16,8 @@
 
 package com.flow.platform.api.dao;
 
+import static com.flow.platform.core.dao.PageUtil.buildPage;
+
 import com.flow.platform.api.domain.Artifact;
 import com.flow.platform.core.dao.AbstractBaseDao;
 import com.flow.platform.core.domain.Page;
