+++ platform-api/src/main/java/com/flow/platform/api/dao/user/UserFlowDaoImpl.java
@@ -15,7 +15,7 @@
  */
 package com.flow.platform.api.dao.user;
 
-import com.flow.platform.api.dao.util.PageUtil;
+import com.flow.platform.core.dao.PageUtil;
 import com.flow.platform.api.domain.user.UserFlow;
 import com.flow.platform.api.domain.user.UserFlowKey;
 import com.flow.platform.core.dao.AbstractBaseDao;
