+++ platform-api/src/main/java/com/flow/platform/api/dao/user/UserRoleDaoImpl.java
@@ -15,7 +15,7 @@
  */
 package com.flow.platform.api.dao.user;
 
-import com.flow.platform.api.dao.util.PageUtil;
+import com.flow.platform.core.dao.PageUtil;
 import com.flow.platform.api.domain.user.UserRole;
 import com.flow.platform.api.domain.user.UserRoleKey;
 import com.flow.platform.core.dao.AbstractBaseDao;
