+++ platform-core/src/main/java/com/flow/platform/core/dao/AbstractBaseDao.java
@@ -157,19 +157,13 @@ public abstract class AbstractBaseDao<K extends Serializable, T> implements Base
             CriteriaQuery<T> select = builder.createQuery(getEntityClass());
             select.from(getEntityClass());
             TypedQuery query = session.createQuery(select);
-            return buildPage(query, pageable, () -> {
+            return PageUtil.buildPage(query, pageable, () -> {
                 CriteriaQuery<Long> selectLong = builder.createQuery(Long.class);
                 Root root = selectLong.from(getEntityClass());
                 selectLong.select(builder.count(root));
                 return session.createQuery(selectLong).uniqueResult();
             });
         });
-    }
-
-    public Page<T> buildPage(TypedQuery query, Pageable pageable, TotalSupplier totalSupplier) {
-        query.setFirstResult(pageable.getOffset());
-        query.setMaxResults(pageable.getPageSize());
-        return new Page<T>(query.getResultList(), pageable.getPageSize(), pageable.getPageNumber(), totalSupplier);
     }
 
     public interface TotalSupplier {
